// functions/src/models/tricy/tricy-interface.ts
export interface TricyModel {
  id: string;
  name: string;
  barangay: string;
  civilStatus: string;
  reason: string;
  make: string;
  motorNo: string;
  chassisNo: string;
  plateNo: string;
  noOfUnits: string;
  franchiseNo: string;
  taxCert: string;
  orNo: string;
  dateFranchise: string;
  year: string;
  status: string;
  fee: string;
  created: string;
  filePath: string;
}